requirements.txt
langchain
langchain-google-genai
langchain-tavily
langchain-weaviate
langchain-community
langgraph
beautifulsoup4
requests
python-dotenv
fastapi
tavily-python
uvicorn[standard]
